refactor: move ValueSequence class methods to a separate module
